Working with Database

+++ Dictionary/Models/Definition.cs
@@ -1,8 +1,10 @@
+using System;
+
 namespace Dictionary.Models
 {
     public class Definition
     {
-        public int Id { get; set; }
+        public Int64 Id { get; set; }
         public string EnglishWord { get; set; }
         public string PartOfSpeech { get; set; }
         public string MalayalamDefinition { get; set; }
